Address review feedback
* revert changes in default primary key strategy
* only change is to pass migration job_arguments to strategies so do
  don't have to hardcode specific group id in the migration strategy
* move traversal ids update inside transaction
* fixed locking of project records
* added clean of namespace gin indexes

+++ db/post_migrate/20211209103048_backfill_project_namespaces_for_group.rb
@@ -8,7 +8,7 @@ class BackfillProjectNamespacesForGroup < Gitlab::Database::Migration[1.0]
   disable_ddl_transaction!
 
   def up
-    return unless Gitlab.com?
+    return unless Gitlab.com? || Gitlab.staging?
 
     projects_table = ::Gitlab::BackgroundMigration::ProjectNamespaces::Models::Project.arel_table
     hierarchy_cte_sql = Arel.sql(::Gitlab::BackgroundMigration::ProjectNamespaces::BackfillProjectNamespaces.hierarchy_cte(GROUP_ID))
@@ -34,7 +34,7 @@ class BackfillProjectNamespacesForGroup < Gitlab::Database::Migration[1.0]
   end
 
   def down
-    return unless Gitlab.com?
+    return unless Gitlab.com? || Gitlab.staging?
 
     Gitlab::Database::BackgroundMigration::BatchedMigration
       .for_configuration(MIGRATION, :projects, :id, [GROUP_ID, 'up']).delete_all
